// frontend/src/service/api/models/FindDmDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FindDmDto = {
    /**
     * Target user ID to find DM with
     */
    userId: number;
    /**
     * Target user ID to find DM with
     */
    targetUserId: number;
};